functions/scripts/python/requirements.txt
python-dotenv>=1.0.0
openai>=1.61.1
langgraph>=0.2.70
langchain-core==0.3.34
langchain-openai==0.3.4
pydantic==2.10.6
